Move functions out of shared lib
