+++ packages/mds-api-helpers/index.ts
@@ -16,162 +16,6 @@
 
 import urls from 'url'
 import express from 'express'
-import { isInsideBoundingBox } from 'mds-utils'
-import {
-  VehicleEvent,
-  Device,
-  Telemetry,
-  BoundingBox,
-  EVENT_STATUS_MAP,
-  VEHICLE_STATUSES,
-  VEHICLE_EVENT
-} from 'mds-types'
-import log from 'mds-logger'
-import db from 'mds-db'
-import cache from 'mds-cache'
-import { CacheReadDeviceResult } from 'mds-cache/types'
-
-export async function getVehiclesDb(
-  skip: number,
-  take: number,
-  url: string,
-  provider_id: string,
-  reqQuery: { [x: string]: string },
-  bbox?: BoundingBox
-): Promise<{
-  total: number
-  links: { first: string; last: string; prev: string | null; next: string | null }
-  vehicles: (Device & { updated?: number | null; telemetry?: Telemetry | null })[]
-}> {
-  function fmt(query: { skip: number; take: number }): string {
-    const flat = Object.assign({}, reqQuery, query)
-    let s = `${url}?`
-    s += Object.keys(flat)
-      .map(key => `${key}=${flat[key]}`)
-      .join('&')
-    return s
-  }
-
-  const rows = await db.readDeviceIds(provider_id)
-  const total = rows.length
-  log.info(`read ${total} deviceIds in /vehicles`)
-
-  const events = await cache.readEvents(rows.map(record => record.device_id))
-  const eventMap: { [s: string]: VehicleEvent } = {}
-  events.map(event => {
-    if (event) {
-      eventMap[event.device_id] = event
-    }
-  })
-
-  const deviceIdSuperset = bbox
-    ? rows.filter(record => {
-        return eventMap[record.device_id] ? isInsideBoundingBox(eventMap[record.device_id].telemetry, bbox) : true
-      })
-    : rows
-
-  const deviceIdSubset = deviceIdSuperset.slice(skip, skip + take).map(record => record.device_id)
-  const devices = (await cache.readDevices(deviceIdSubset)).reduce((acc: Device[], device: CacheReadDeviceResult) => {
-    if (!device) {
-      throw new Error('device in DB but not in cache')
-    }
-    const event = eventMap[device.device_id]
-    const status = event ? EVENT_STATUS_MAP[event.event_type] : VEHICLE_STATUSES.removed
-    const telemetry = event ? event.telemetry : null
-    const updated = event ? event.timestamp : null
-    return [...acc, { ...device, status, telemetry, updated }]
-  }, [])
-
-  const noNext = skip + take >= deviceIdSuperset.length
-  const noPrev = skip === 0 || skip > deviceIdSuperset.length
-  const lastSkip = take * Math.floor(deviceIdSuperset.length / take)
-
-  return {
-    total,
-    links: {
-      first: fmt({
-        skip: 0,
-        take
-      }),
-      last: fmt({
-        skip: lastSkip,
-        take
-      }),
-      prev: noPrev
-        ? null
-        : fmt({
-            skip: skip - take,
-            take
-          }),
-      next: noNext
-        ? null
-        : fmt({
-            skip: skip + take,
-            take
-          })
-    },
-    vehicles: devices
-  }
-}
-
-export async function getVehiclesCache(
-  skip: number,
-  take: number,
-  url: string,
-  provider_id: string,
-  reqQuery: { [x: string]: string },
-  bbox: BoundingBox
-) {
-  function fmt(query: { skip: number; take: number }): string {
-    const flat = Object.assign({}, reqQuery, query)
-    let s = `${url}?`
-    s += Object.keys(flat)
-      .map(key => `${key}=${flat[key]}`)
-      .join('&')
-    return s
-  }
-
-  const statusesSuperset = ((await cache.readDevicesStatus({ bbox })) as (VehicleEvent & Device)[]).filter(
-    status => EVENT_STATUS_MAP[status.event_type as VEHICLE_EVENT] !== VEHICLE_STATUSES.removed
-  )
-  const statusesSubset = statusesSuperset.slice(skip, skip + take)
-  const devices = statusesSubset.reduce((acc: (VehicleEvent & Device)[], item) => {
-    const status = EVENT_STATUS_MAP[item.event_type as VEHICLE_EVENT]
-    const updated = item.timestamp
-    return [...acc, { ...item, status, updated }]
-  }, [])
-
-  const noNext = skip + take >= statusesSuperset.length
-  const noPrev = skip === 0 || skip > statusesSuperset.length
-  const lastSkip = take * Math.floor(statusesSuperset.length / take)
-
-  return {
-    total: statusesSuperset.length,
-    links: {
-      first: fmt({
-        skip: 0,
-        take
-      }),
-      last: fmt({
-        skip: lastSkip,
-        take
-      }),
-      prev: noPrev
-        ? null
-        : fmt({
-            skip: skip - take,
-            take
-          }),
-      next: noNext
-        ? null
-        : fmt({
-            skip: skip + take,
-            take
-          })
-    },
-    vehicles: devices
-  }
-}
 
 interface PagingParams {
   skip: number
